% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{alpha_diversity}
\alias{alpha_diversity}
\title{Returns alpha diversity measures and pvalues}
\usage{
alpha_diversity(ps_rare)
}
\arguments{
\item{ps_rare}{- phyloseq object to calculate alpha diversity on}
}
\value{

}
\description{
Given a ps object, calculate alpha diversity on it and run a t-test
on group
}
