% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{diff_analysis_lefse}
\alias{diff_analysis_lefse}
\title{Run lefse analysis on phyloseq object with cutoffs}
\usage{
diff_analysis_lefse(ps_new, secondalpha = 0.05, firstalpha = 0.1)
}
\arguments{
\item{secondalpha}{}
}
\value{

}
\description{
Run lefse analysis on phyloseq object with cutoffs
}
