% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{filter_and_plot_correlations}
\alias{filter_and_plot_correlations}
\title{Filter correlations based on cutoff and plot them}
\usage{
filter_and_plot_correlations(
  basdai_iga_corr,
  pvalue_column = pvalue_iga,
  variable = IgA_index,
  alpha = 0.05,
  level = Phylum_Genus_ASV
)
}
\arguments{
\item{basdai_iga_corr}{- correlation output from \code{run_basdai_correlations}}

\item{level}{}
}
\value{

}
\description{
Filter correlations based on cutoff and plot them
}
