% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{keep_levels}
\alias{keep_levels}
\title{Filters lefse analysis to only contain one level of analysis}
\usage{
keep_levels(diffres, level = "Species")
}
\arguments{
\item{diffres}{- differential results object processed
with @code{rewrite_annotation}}

\item{level}{- level of analysis to retain - default is Species}
}
\value{

}
\description{
Filters lefse analysis to only contain one level of analysis
}
