% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{lefse_plot}
\alias{lefse_plot}
\title{Returns a plot of lefse candidates - abundances separated by group}
\usage{
lefse_plot(diffres)
}
\arguments{
\item{diffres}{}
}
\value{

}
\description{
Returns a plot of lefse candidates - abundances separated by group
}
