% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{rewrite_annotation}
\alias{rewrite_annotation}
\title{Removes confusing multi level annotations from lefse analysis}
\usage{
rewrite_annotation(diffres)
}
\arguments{
\item{diffres}{- differential results analysis from diff_analysis_lefse}
}
\value{

}
\description{
Removes confusing multi level annotations from lefse analysis
}
