% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{run_basdai_alpha_correlations}
\alias{run_basdai_alpha_correlations}
\title{Correlate iga correlations}
\usage{
run_basdai_alpha_correlations(
  iga_basdai_alpha,
  level = Phylum_Genus_ASV,
  method = "spearman"
)
}
\arguments{
\item{iga_basdai_alpha}{- IgA Index calculations to correlate}

\item{level}{- level to calculate correlations on}

\item{method}{- correlation method to use - passed on to @code{cor.test}.
Currently "spearman" (default) or "pearson".}
}
\value{

}
\description{
Correlate iga correlations
}
