% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{run_basdai_iga_correlations}
\alias{run_basdai_iga_correlations}
\title{Given alpha values and basdai}
\usage{
run_basdai_iga_correlations(
  iga_basdai,
  level = Phylum_Genus_ASV,
  method = "spearman"
)
}
\arguments{
\item{level}{}
}
\value{

}
\description{
Given alpha values and basdai
}
