% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{calc_alpha_diversity}
\alias{calc_alpha_diversity}
\title{Calculates alpha diversity using three calculations}
\usage{
calc_alpha_diversity(ps_rare, subject_id = subject_id)
}
\arguments{
\item{subject_id}{}
}
\value{

}
\description{
Calculates alpha diversity using three calculations
}
