% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{diff_analysis_lefse}
\alias{diff_analysis_lefse}
\title{Title}
\usage{
diff_analysis_lefse(ps_new, secondalpha = 0.05)
}
\arguments{
\item{secondalpha}{}
}
\value{

}
\description{
Title
}
